requirements.txt
kafka-python==2.0.2
pandera==0.20.4
pandas==2.1.4
